Fix API listFiles.php structure api and mappers encapsulate mapModified(working on)

+++ api/removeElement.php
@@ -1,30 +1,19 @@
 <?php
 header("access-control-allow-origin: http://localhost:3000");
+header('Content-Type: application/json');
 
-$absolutePath = 'C:\Users\Nayxer\Projects\Atom\root';
-$dir = $absolutePath.'\\'.$_POST['folderName'];
+$absolutePath = '/home/snowtray/test/';
+$dir = $absolutePath.'/'.$_POST['folderName'];
 
 function deleteDirectory($dir) {
-    if (!file_exists($dir)) {
-        return true;
-    }
-
-    if (!is_dir($dir)) {
-        return unlink($dir);
-    }
+    if (!file_exists($dir)) return true;
+    if (!is_dir($dir)) return unlink($dir);
 
     foreach (scandir($dir) as $item) {
-        if ($item == '.' || $item == '..') {
-            continue;
-        }
-
-        if (!deleteDirectory($dir . DIRECTORY_SEPARATOR . $item)) {
-            return false;
-        }
-
+        if ($item == '.' || $item == '..') continue;
+        if (!deleteDirectory($dir . DIRECTORY_SEPARATOR . $item)) return false;
     }
-
     return rmdir($dir);
 }
 
-deleteDirectory($dir);
+deleteDirectory($dir)?'true':'false';
